Restructure the application
