+++ app.py
@@ -1,4 +1,4 @@
-from . import app
+from flask_app import app
 
 
 if __name__ == "__main__":
