+++ flask_app/models/therapist.py
@@ -1,4 +1,4 @@
-from app import db
+from flask_app import db
 
 
 class Therapist(db.Model):
@@ -8,6 +8,7 @@ class Therapist(db.Model):
     image_profile = db.Column(db.String(20), nullable=False, default='default.jpg')
     password = db.Column(db.String(60), nullable=False)
     specialization = db.Column(db.String(100), nullable=False)
+    license_number = db.Column(db.Integer, nullable=False, unique=True)
     
     appointments = db.relationship('Appointment', backref='doctor', lazy=True)
