+++ flask_app/models/user.py
@@ -1,4 +1,4 @@
-from app import db
+from flask_app import db
 from sqlalchemy.sql import func
