+++ flask_app/templates/register_therapist.html
