--- forms.py
@@ -1,34 +1,0 @@
-from flask_wtf import FlaskForm
-from wtforms import StringField, SubmitField, PasswordField, BooleanField
-from wtforms.validators import DataRequired, Email, EqualTo, Length, ValidationError
-
-
-class RegistrationForm(FlaskForm):
-    username = StringField('Username', validators=[DataRequired()])
-    email = StringField('Email', validators=[DataRequired(), Email()])
-    password = PasswordField('Password', validators=[DataRequired()])
-    confirm_password = PasswordField( 'Confirm Password', validators=[DataRequired(), EqualTo('password')])
-    submit = SubmitField('Sign Up')
-
-
-class LoginForm(FlaskForm):
-    email = StringField('Email', validators=[DataRequired(), Email()])
-    password = PasswordField('Password', validators=[DataRequired()])
-    remember = BooleanField('Remember Me')
-    submit = SubmitField('Log In')
-
-class TherapistRegistrationForm(FlaskForm):
-        first_name = StringField('First Name', validators=[DataRequired()])
-        last_name = StringField('Last Name', validators=[DataRequired()])
-        username = StringField('Username', validators=[DataRequired()])
-        email = StringField('Email', validators=[DataRequired(), Email()])
-        password = PasswordField('Password', validators=[DataRequired()])
-        confirm_password = PasswordField('Confirm Password', validators=[DataRequired(), EqualTo('password')])
-        license_number = StringField('License Number', validators=[DataRequired(), Length(min=6, max=12)])
-        submit = SubmitField('Register')
-
-class TherapistLoginForm(FlaskForm):
-        email = StringField('Email', validators=[DataRequired(), Email()])
-        password = PasswordField('Password', validators=[DataRequired()])
-        remember = BooleanField('Remember Me')
-        submit = SubmitField('Login')
